Extracted ContactAddress out to its own view
